# openapi.yaml
openapi: 3.0.3
info:
  title: Shipping Service Orchestration API
  description: >
    Comprehensive OpenAPI 3.0 specification for the Shipping Microservices Ecosystem.
    This document covers endpoints across all microservices: Orchestrator, Customer Service,
    Shipment Service, Payment Service, and Tracking Service.
  version: 1.0.0
  contact:
    name: Engineering Team

servers:
  - url: http://localhost:8080
    description: API Gateway (Unified Routing Entry point)
  - url: http://localhost:8081
    description: Orchestrator Service (Port 8081)
  - url: http://localhost:8084
    description: Customer Service (Port 8084)
  - url: http://localhost:8083
    description: Shipment Service (Port 8083)
  - url: http://localhost:8082
    description: Payment Service (Port 8082)
  - url: http://localhost:8086
    description: Tracking Service (Port 8086)

tags:
  - name: Orchestrator
    description: Order Creation Saga & Orchestration Endpoints
  - name: Customer Service
    description: Customer Profile Management
  - name: Shipment Service
    description: Shipment Operations & Logistics Pricing
  - name: Payment Service
    description: Payment Processing & Audit Records
  - name: Tracking Service
    description: Package Tracking & History Updates

paths:
  # ==========================================
  # ORCHESTRATOR SERVICE ENDPOINTS
  # ==========================================
  /api/orders:
    post:
      tags:
        - Orchestrator
      summary: Create & Orchestrate Shipping Order (Saga Flow)
      description: >
        Triggers the multi-service Saga workflow:
        1. Validates customer via Customer Service.
        2. Creates shipment (PENDING) and calculates fee via Shipment Service.
        3. Processes payment via Payment Service. If payment fails, compensates by deleting shipment.
        4. Confirms shipment status to CONFIRMED.
        5. Registers tracking record via Tracking Service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerId: 1
              originCity: "Jakarta"
              destinationCity: "Surabaya"
              weight: 2.5
              serviceType: "EXPRESS"
              paymentMethod: "CREDIT_CARD"
      responses:
        '201':
          description: Order created successfully with COMPLETED status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'
        '422':
          description: Saga execution failed or unprocessable (Compensating actions executed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'

    get:
      tags:
        - Orchestrator
      summary: Get All Orchestrated Orders (In-Memory Log)
      description: Returns list of order saga outcomes currently stored in the Orchestrator in-memory cache.
      responses:
        '200':
          description: List of orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResult'

  /api/orders/{id}:
    get:
      tags:
        - Orchestrator
      summary: Get Order Outcome by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'
        '404':
          description: Order not found

  # ==========================================
  # CUSTOMER SERVICE ENDPOINTS
  # ==========================================
  /api/customers:
    post:
      tags:
        - Customer Service
      summary: Create New Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
            example:
              name: "John Doe"
              email: "john.doe@example.com"
              phoneNumber: "08123456789"
              address: "Jl. Sudirman No. 45, Jakarta"
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

    get:
      tags:
        - Customer Service
      summary: Get All Customers
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'

  /api/customers/{id}:
    get:
      tags:
        - Customer Service
      summary: Get Customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found

    put:
      tags:
        - Customer Service
      summary: Update Customer Details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found

    delete:
      tags:
        - Customer Service
      summary: Delete Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Customer deleted
        '500':
          description: Server Error

  # ==========================================
  # SHIPMENT SERVICE ENDPOINTS
  # ==========================================
  /api/shipments:
    post:
      tags:
        - Shipment Service
      summary: Create Shipment & Calculate Fee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shipment'
            example:
              customerId: 1
              originCity: "Jakarta"
              destinationCity: "Surabaya"
              weight: 2.5
              serviceType: "EXPRESS"
              status: "PENDING"
      responses:
        '201':
          description: Shipment created with auto-calculated pricing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '400':
          description: Pricing data missing for given origin, destination, and service type

    get:
      tags:
        - Shipment Service
      summary: Get All Shipments
      responses:
        '200':
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shipment'

  /api/shipments/{id}:
    get:
      tags:
        - Shipment Service
      summary: Get Shipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '404':
          description: Shipment not found

    put:
      tags:
        - Shipment Service
      summary: Update Shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shipment'
      responses:
        '200':
          description: Shipment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '404':
          description: Shipment not found

    delete:
      tags:
        - Shipment Service
      summary: Delete / Rollback Shipment (Compensating action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: Shipment deleted successfully

  /api/shipments/track/{trackingNumber}:
    get:
      tags:
        - Shipment Service
      summary: Track Shipment by Tracking Number
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
          example: "RESI-1712345678900"
      responses:
        '200':
          description: Shipment detail by tracking number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '404':
          description: Tracking number not found

  /api/shipments/pricing:
    post:
      tags:
        - Shipment Service
      summary: Create Pricing Rate Master Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pricing'
            example:
              originCity: "Jakarta"
              destinationCity: "Surabaya"
              serviceType: "EXPRESS"
              price: 15000.00
              weight: 1.0
      responses:
        '201':
          description: Pricing rate added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pricing'

    get:
      tags:
        - Shipment Service
      summary: Get All Pricing Master Rates
      responses:
        '200':
          description: List of pricing rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pricing'

  /api/shipments/pricing/{id}:
    get:
      tags:
        - Shipment Service
      summary: Get Pricing Rate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Pricing rate details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pricing'
        '404':
          description: Pricing rate not found

    put:
      tags:
        - Shipment Service
      summary: Update Pricing Rate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pricing'
      responses:
        '200':
          description: Pricing updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pricing'

    delete:
      tags:
        - Shipment Service
      summary: Delete Pricing Rate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Pricing rate deleted

  # ==========================================
  # PAYMENT SERVICE ENDPOINTS
  # ==========================================
  /api/payments:
    post:
      tags:
        - Payment Service
      summary: Process Payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
            example:
              shipmentId: 1
              amount: 37500.00
              paymentMethod: "CREDIT_CARD"
              paymentStatus: "SUCCESS"
      responses:
        '201':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

    get:
      tags:
        - Payment Service
      summary: Get All Payment Records
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'

  /api/payments/{id}:
    get:
      tags:
        - Payment Service
      summary: Get Payment by Document ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60d5ec49f1b2c81a4c8e4567"
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Payment not found

    put:
      tags:
        - Payment Service
      summary: Update Payment Record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60d5ec49f1b2c81a4c8e4567"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
      responses:
        '200':
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

    delete:
      tags:
        - Payment Service
      summary: Delete Payment Record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60d5ec49f1b2c81a4c8e4567"
      responses:
        '204':
          description: Payment record deleted

  /api/payments/shipment/{shipmentId}:
    get:
      tags:
        - Payment Service
      summary: Get Payment by Shipment ID
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Payment details for shipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Payment for shipment not found

  # ==========================================
  # TRACKING SERVICE ENDPOINTS
  # ==========================================
  /api/tracking:
    post:
      tags:
        - Tracking Service
      summary: Initiate Package Tracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tracking'
            example:
              trackingNumber: "RESI-1712345678900"
              currentStatus: "SHIPMENT_CREATED"
              history:
                - status: "SHIPMENT_CREATED"
                  location: "Jakarta Origin Hub"
                  timestamp: "2026-07-23T00:00:00"
      responses:
        '201':
          description: Tracking document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'

  /api/tracking/{trackingNumber}:
    get:
      tags:
        - Tracking Service
      summary: Get Package Tracking History by Resi
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
          example: "RESI-1712345678900"
      responses:
        '200':
          description: Tracking history details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
        '404':
          description: Tracking number not found

    delete:
      tags:
        - Tracking Service
      summary: Delete Tracking Record
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
          example: "RESI-1712345678900"
      responses:
        '204':
          description: Tracking record deleted

  /api/tracking/{trackingNumber}/update-status:
    put:
      tags:
        - Tracking Service
      summary: Update Tracking Status & Push History Event
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
          example: "RESI-1712345678900"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackingHistory'
            example:
              status: "IN_TRANSIT"
              location: "Bandung Sorting Center"
              timestamp: "2026-07-23T04:30:00"
      responses:
        '200':
          description: Tracking history appended and status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'

components:
  schemas:
    CreateOrderRequest:
      type: object
      properties:
        customerId:
          type: integer
          example: 1
        originCity:
          type: string
          example: "Jakarta"
        destinationCity:
          type: string
          example: "Surabaya"
        weight:
          type: number
          format: double
          example: 2.5
        serviceType:
          type: string
          example: "EXPRESS"
        paymentMethod:
          type: string
          example: "CREDIT_CARD"

    OrderResult:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          example: 1001
        customerId:
          type: integer
          example: 1
        originCity:
          type: string
          example: "Jakarta"
        destinationCity:
          type: string
          example: "Surabaya"
        weight:
          type: number
          format: double
          example: 2.5
        serviceType:
          type: string
          example: "EXPRESS"
        paymentMethod:
          type: string
          example: "CREDIT_CARD"
        shipmentId:
          type: integer
          example: 50
        trackingNumber:
          type: string
          example: "RESI-1712345678900"
        price:
          type: number
          example: 37500.00
        paymentId:
          type: string
          example: "60d5ec49f1b2c81a4c8e4567"
        status:
          type: string
          example: "COMPLETED"
        failureReason:
          type: string
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "john.doe@example.com"
        phoneNumber:
          type: string
          example: "08123456789"
        address:
          type: string
          example: "Jl. Sudirman No. 45, Jakarta"

    Shipment:
      type: object
      properties:
        id:
          type: integer
          example: 50
        trackingNumber:
          type: string
          example: "RESI-1712345678900"
        customerId:
          type: integer
          example: 1
        originCity:
          type: string
          example: "Jakarta"
        destinationCity:
          type: string
          example: "Surabaya"
        weight:
          type: number
          format: double
          example: 2.5
        serviceType:
          type: string
          example: "EXPRESS"
        price:
          type: number
          example: 37500.00
        status:
          type: string
          example: "CONFIRMED"

    Pricing:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        originCity:
          type: string
          example: "Jakarta"
        destinationCity:
          type: string
          example: "Surabaya"
        serviceType:
          type: string
          example: "EXPRESS"
        price:
          type: number
          example: 15000.00
        weight:
          type: number
          format: double
          example: 1.0

    Payment:
      type: object
      properties:
        id:
          type: string
          example: "60d5ec49f1b2c81a4c8e4567"
        shipmentId:
          type: integer
          example: 50
        amount:
          type: number
          example: 37500.00
        paymentMethod:
          type: string
          example: "CREDIT_CARD"
        paymentStatus:
          type: string
          example: "SUCCESS"
        paymentDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Tracking:
      type: object
      properties:
        id:
          type: string
          example: "60d5ec49f1b2c81a4c8e4568"
        trackingNumber:
          type: string
          example: "RESI-1712345678900"
        currentStatus:
          type: string
          example: "IN_TRANSIT"
        history:
          type: array
          items:
            $ref: '#/components/schemas/TrackingHistory'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TrackingHistory:
      type: object
      properties:
        status:
          type: string
          example: "IN_TRANSIT"
        location:
          type: string
          example: "Bandung Sorting Center"
        timestamp:
          type: string
          format: date-time